refactoring kcfg show to show proof of a claim
