+++ demo/SIMPLE-SPEC.statement.1/kcfg/kcfg.json
@@ -1,0 +1,1 @@
+{"next": 4, "nodes": [1, 2, 3], "edges": [{"source": 1, "target": 3, "depth": 1}], "covers": [{"source": 3, "target": 2, "csubst": {"subst": {"GENERATEDCOUNTER_CELL_5e3e01ba": {"node": "KVariable", "name": "GENERATEDCOUNTER_CELL_949ec677", "sort": {"node": "KSort", "name": "Int"}}}, "constraints": [{"node": "KApply", "label": {"node": "KLabel", "name": "#Equals", "params": [{"node": "KSort", "name": "Bool"}, {"node": "KSort", "name": "GeneratedTopCell"}]}, "args": [{"node": "KToken", "token": "false", "sort": {"node": "KSort", "name": "Bool"}}, {"node": "KApply", "label": {"node": "KLabel", "name": "LocalDeclCellMap:in_keys", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "<index>", "params": []}, "args": [{"node": "KToken", "token": "_3", "sort": {"node": "KSort", "name": "LocalToken"}}], "arity": 1, "variable": false}, {"node": "KVariable", "name": "LOCALDECLS_CELL_34597eca", "sort": {"node": "KSort", "name": "LocalDeclCellMap"}}], "arity": 2, "variable": false}], "arity": 2, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "#Equals", "params": [{"node": "KSort", "name": "Bool"}, {"node": "KSort", "name": "GeneratedTopCell"}]}, "args": [{"node": "KToken", "token": "false", "sort": {"node": "KSort", "name": "Bool"}}, {"node": "KApply", "label": {"node": "KLabel", "name": "FunctionCellMap:in_keys", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "<fnKey>", "params": []}, "args": [{"node": "KVariable", "name": "FN_KEY", "sort": {"node": "KSort", "name": "FunctionLikeKey"}}], "arity": 1, "variable": false}, {"node": "KVariable", "name": "FUNCTIONS_CELL_fe1cac3f", "sort": {"node": "KSort", "name": "FunctionCellMap"}}], "arity": 2, "variable": false}], "arity": 2, "variable": false}]}}], "splits": [], "ndbranches": [], "vacuous": [], "stuck": [], "aliases": {}}
