+++ demo/SIMPLE-SPEC.statement.1/kcfg/nodes/1.json
@@ -1,0 +1,1 @@
+{"id": 1, "cterm": {"config": {"node": "KApply", "label": {"node": "KLabel", "name": "<generatedTop>", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "<k>", "params": []}, "args": [{"node": "KSequence", "items": [{"node": "KApply", "label": {"node": "KLabel", "name": "#executeStatement(_)_MIR-EXECUTION_MirSimulation_StatementKind", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "_=__MIR-STATEMENT-SYNTAX_Assign_Place_RValue", "params": []}, "args": [{"node": "KToken", "token": "_3", "sort": {"node": "KSort", "name": "LocalToken"}}, {"node": "KApply", "label": {"node": "KLabel", "name": "const__MIR-CONSTANT-SYNTAX_Constant_ConstantValue", "params": []}, "args": [{"node": "KToken", "token": "5_usize", "sort": {"node": "KSort", "name": "UnsignedLiteral"}}], "arity": 1, "variable": false}], "arity": 2, "variable": false}], "arity": 1, "variable": false}, {"node": "KVariable", "name": "K_CELL_fc656f08", "sort": {"node": "KSort", "name": "K"}}], "arity": 2}], "arity": 1, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "<returncode>", "params": []}, "args": [{"node": "KVariable", "name": "RETURNCODE_CELL", "sort": {"node": "KSort", "name": "Int"}}], "arity": 1, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "<mir>", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "<simulator>", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "<callStack>", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "_List_", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "ListItem", "params": []}, "args": [{"node": "KVariable", "name": "FN_KEY", "sort": {"node": "KSort", "name": "FunctionLikeKey"}}], "arity": 1, "variable": false}, {"node": "KVariable", "name": "CALLSTACK_CELL_c2f4a5db", "sort": {"node": "KSort", "name": "List"}}], "arity": 2, "variable": false}], "arity": 1, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "<currentBasicBlock>", "params": []}, "args": [{"node": "KVariable", "name": "CURRENTBASICBLOCK_CELL", "sort": {"node": "KSort", "name": "Int"}}], "arity": 1, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "<phase>", "params": []}, "args": [{"node": "KVariable", "name": "PHASE_CELL", "sort": {"node": "KSort", "name": "MirSimulatorPhase"}}], "arity": 1, "variable": false}], "arity": 3, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "<data>", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "<functions>", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "_FunctionCellMap_", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "<function>", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "<fnKey>", "params": []}, "args": [{"node": "KVariable", "name": "FN_KEY", "sort": {"node": "KSort", "name": "FunctionLikeKey"}}], "arity": 1, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "<localDecls>", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "_LocalDeclCellMap_", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "<localDecl>", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "<index>", "params": []}, "args": [{"node": "KToken", "token": "_3", "sort": {"node": "KSort", "name": "LocalToken"}}], "arity": 1, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "<mutability>", "params": []}, "args": [{"node": "KVariable", "name": "MUTABILITY_CELL", "sort": {"node": "KSort", "name": "Mutability"}}], "arity": 1, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "<internal>", "params": []}, "args": [{"node": "KVariable", "name": "INTERNAL_CELL", "sort": {"node": "KSort", "name": "Bool"}}], "arity": 1, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "<ty>", "params": []}, "args": [{"node": "KVariable", "name": "TY_CELL", "sort": {"node": "KSort", "name": "Type"}}], "arity": 1, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "<value>", "params": []}, "args": [{"node": "KVariable", "name": "VALUE_CELL_ed4098dd", "sort": {"node": "KSort", "name": "RValueResult"}}], "arity": 1, "variable": false}], "arity": 5, "variable": false}, {"node": "KVariable", "name": "LOCALDECLS_CELL_34597eca", "sort": {"node": "KSort", "name": "LocalDeclCellMap"}}], "arity": 2, "variable": false}], "arity": 1, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "<basicBlocks>", "params": []}, "args": [{"node": "KVariable", "name": "BASICBLOCKS_CELL", "sort": {"node": "KSort", "name": "BasicBlockCellMap"}}], "arity": 1, "variable": false}], "arity": 3, "variable": false}, {"node": "KVariable", "name": "FUNCTIONS_CELL_fe1cac3f", "sort": {"node": "KSort", "name": "FunctionCellMap"}}], "arity": 2, "variable": false}], "arity": 1, "variable": false}], "arity": 1, "variable": false}], "arity": 2, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "<generatedCounter>", "params": []}, "args": [{"node": "KVariable", "name": "GENERATEDCOUNTER_CELL_949ec677", "sort": {"node": "KSort", "name": "Int"}}], "arity": 1, "variable": false}], "arity": 4, "variable": false}, "constraints": [{"node": "KApply", "label": {"node": "KLabel", "name": "#Equals", "params": [{"node": "KSort", "name": "Bool"}, {"node": "KSort", "name": "GeneratedTopCell"}]}, "args": [{"node": "KToken", "token": "false", "sort": {"node": "KSort", "name": "Bool"}}, {"node": "KApply", "label": {"node": "KLabel", "name": "LocalDeclCellMap:in_keys", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "<index>", "params": []}, "args": [{"node": "KToken", "token": "_3", "sort": {"node": "KSort", "name": "LocalToken"}}], "arity": 1, "variable": false}, {"node": "KVariable", "name": "LOCALDECLS_CELL_34597eca", "sort": {"node": "KSort", "name": "LocalDeclCellMap"}}], "arity": 2, "variable": false}], "arity": 2, "variable": false}, {"node": "KApply", "label": {"node": "KLabel", "name": "#Equals", "params": [{"node": "KSort", "name": "Bool"}, {"node": "KSort", "name": "GeneratedTopCell"}]}, "args": [{"node": "KToken", "token": "false", "sort": {"node": "KSort", "name": "Bool"}}, {"node": "KApply", "label": {"node": "KLabel", "name": "FunctionCellMap:in_keys", "params": []}, "args": [{"node": "KApply", "label": {"node": "KLabel", "name": "<fnKey>", "params": []}, "args": [{"node": "KVariable", "name": "FN_KEY", "sort": {"node": "KSort", "name": "FunctionLikeKey"}}], "arity": 1, "variable": false}, {"node": "KVariable", "name": "FUNCTIONS_CELL_fe1cac3f", "sort": {"node": "KSort", "name": "FunctionCellMap"}}], "arity": 2, "variable": false}], "arity": 2, "variable": false}]}}
