+++ demo/SIMPLE-SPEC.statement.1/proof.json
@@ -1,0 +1,14 @@
+{
+    "id": "SIMPLE-SPEC.statement.1",
+    "subproof_ids": [],
+    "admitted": false,
+    "type": "APRProof",
+    "init": 1,
+    "target": 2,
+    "terminal": [],
+    "node_refutations": {},
+    "circularity": false,
+    "logs": {
+        "3": []
+    }
+}
